+++ lib/scala/runtime-version-manager-test/src/main/scala/org/enso/runtimeversionmanager/test/SlowTestSynchronizer.scala
@@ -1,0 +1,7 @@
+package org.enso.runtimeversionmanager.test
+
+import org.enso.testkit.TestSynchronizer
+
+class SlowTestSynchronizer extends TestSynchronizer {
+  override val timeOutSeconds: Long = 90
+}
