+++ lib/scala/runtime-version-manager-test/src/test/scala/org/enso/distribution/locking/ConcurrencyTest.scala
@@ -22,12 +22,7 @@ import org.enso.runtimeversionmanager.releases.engine.{
 import org.enso.runtimeversionmanager.releases.graalvm.GraalCEReleaseProvider
 import org.enso.runtimeversionmanager.releases.testing.FakeReleaseProvider
 import org.enso.runtimeversionmanager.test._
-import org.enso.testkit.{
-  FlakySpec,
-  RetrySpec,
-  TestSynchronizer,
-  WithTemporaryDirectory
-}
+import org.enso.testkit.{FlakySpec, RetrySpec, WithTemporaryDirectory}
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.concurrent.TimeLimitedTests
 import org.scalatest.matchers.should.Matchers
@@ -69,10 +64,6 @@ class ConcurrencyTest
       callback(packageFileName)
       originalRelease.downloadPackage(destination)
     }
-  }
-
-  class SlowTestSynchronizer extends TestSynchronizer {
-    override val timeOutSeconds: Long = 90
   }
 
   var testLocalLockManager: Option[LockManager] = None
@@ -367,73 +358,6 @@ class ConcurrencyTest
         "t2-start-uninstall-and-wait",
         "t1-end-using",
         "t2-uninstalled"
-      )
-    }
-
-    "synchronize main lock" taggedAs Retry in {
-
-      /** First two threads start and acquire the shared lock, than the third
-        * thread tries to acquire an exclusive lock (in practice that will be our
-        * (un)installer), it should wait for the other threads to finish. When
-        * the threads see that it started waiting (the waiting notification is
-        * normally used to tell the user what the application is waiting for),
-        * the two threads finish and after that the third one is able to acquire
-        * the exclusive lock.
-        */
-      val sync = new SlowTestSynchronizer
-      sync.startThread("t1") {
-        val resourceManager = makeNewResourceManager()
-        resourceManager.initializeMainLock()
-        sync.report("shared-start")
-        sync.signal("started-1")
-        sync.waitFor("finish-1")
-        sync.report("shared-end")
-        resourceManager.releaseMainLock()
-      }
-
-      sync.startThread("t2") {
-        val resourceManager = makeNewResourceManager()
-        resourceManager.initializeMainLock()
-        sync.report("shared-start")
-        sync.signal("started-2")
-        sync.waitFor("finish-2")
-        sync.report("shared-end")
-        resourceManager.releaseMainLock()
-      }
-
-      sync.waitFor("started-1")
-      sync.waitFor("started-2")
-
-      sync.startThread("t3") {
-        val resourceManager = makeNewResourceManager()
-        resourceManager.initializeMainLock()
-        sync.report("t3-start")
-        resourceManager.acquireExclusiveMainLock(() => {
-          sync.report("t3-wait")
-          sync.signal("waiting")
-        })
-        sync.report("t3-end")
-        sync.signal("finish-all")
-        resourceManager.releaseMainLock()
-      }
-
-      sync.waitFor("waiting")
-      Thread.sleep(1000)
-
-      sync.signal("finish-1")
-      sync.signal("finish-2")
-
-      sync.waitFor("finish-all")
-
-      sync.join()
-      sync.summarizeReports() shouldEqual Seq(
-        "shared-start",
-        "shared-start",
-        "t3-start",
-        "t3-wait",
-        "shared-end",
-        "shared-end",
-        "t3-end"
       )
     }
   }
